Fix 1: Use stream base time if no prediction base time is set
Fix 2: Convert to double to get correct value for m

Former-commit-id: c574c8a7fbb8618a0dec5574cef4ee1ef191fc47

+++ server/spatiotemporal/de.uniol.inf.is.odysseus.temporaltypes/src/de/uniol/inf/is/odysseus/temporaltypes/function/NonIncrementalTemporalizeIntegerFunction.java
@@ -107,7 +107,7 @@ public class NonIncrementalTemporalizeIntegerFunction<M extends ITimeInterval, T
 		long temporalDistance = Math.abs(trigger.getMetadata().getStart().minus(oldestStartTime).getMainPoint());
 		if (temporalDistance != 0) {
 			// Oldest and newest element are not valid at the same time -> we have a slope
-			m = (Math.abs(newestValue - oldestValue)) / temporalDistance;			
+			m = ((double) Math.abs(newestValue - oldestValue)) / temporalDistance;			
 		}
 		
 		// Calculate y axis intercept
